feat(beam-domain): extract shared domain models and repository traits into new crate
Create a new `beam-domain` workspace crate containing:
- Domain models: Movie, Show, Season, Episode, MediaFile, MediaStream, Library, AdminLog
- Repository traits: LibraryRepository, FileRepository, MovieRepository, ShowRepository,
  MediaStreamRepository, AdminLogRepository
- InMemory implementations gated behind `test-utils` feature
- `From<beam_entity::*>` impls gated behind `entity` feature

Update beam-index to re-export domain models and repository traits from beam-domain,
keeping only SQL implementations locally. Remove duplicated local model files.

Update beam-stream to import domain types directly from beam-domain, and delete
previously orphaned duplicate model and repository files.

This eliminates manual synchronization when the schema changes, as domain models
are now defined once in beam-domain.
