+++ beam-domain/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "beam-domain"
+version = "0.1.0"
+edition = "2024"
+license = "AGPL-3.0-only"
+publish = false
+
+[dependencies]
+async-trait = { workspace = true }
+chrono = { workspace = true }
+parking_lot = "0.12.5"
+sea-orm = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+uuid = { workspace = true }
+
+[dependencies.beam-entity]
+path = "../beam-entity"
+optional = true
+
+[dependencies.mockall]
+version = "0.14.0"
+optional = true
+
+[features]
+entity = ["dep:beam-entity"]
+test-utils = ["dep:mockall"]
