+++ beam-domain/src/models/file.rs
@@ -91,9 +91,9 @@ pub struct UpdateMediaFile {
     pub status: Option<FileStatus>,
 }
 
+#[cfg(feature = "entity")]
 impl From<beam_entity::files::Model> for MediaFile {
     fn from(model: beam_entity::files::Model) -> Self {
-        // Determine content type from polymorphic foreign keys
         let content = model
             .movie_entry_id
             .map(|id| MediaFileContent::Movie { movie_entry_id: id })
