+++ beam-domain/src/models/mod.rs
@@ -1,0 +1,13 @@
+pub mod admin_log;
+pub mod file;
+pub mod library;
+pub mod movie;
+pub mod show;
+pub mod stream;
+
+pub use admin_log::*;
+pub use file::*;
+pub use library::*;
+pub use movie::*;
+pub use show::*;
+pub use stream::*;
