+++ beam-domain/src/models/movie.rs
@@ -1,4 +1,4 @@
-use chrono::{DateTime, Utc};
+use chrono::{DateTime, NaiveDate, Utc};
 use std::time::Duration;
 use uuid::Uuid;
 
@@ -7,7 +7,18 @@ use uuid::Uuid;
 pub struct Movie {
     pub id: Uuid,
     pub title: String,
+    pub title_localized: Option<String>,
+    pub description: Option<String>,
+    pub year: Option<u32>,
+    pub release_date: Option<NaiveDate>,
     pub runtime: Option<Duration>,
+    pub poster_url: Option<String>,
+    pub backdrop_url: Option<String>,
+    pub tmdb_id: Option<u32>,
+    pub imdb_id: Option<String>,
+    pub tvdb_id: Option<u32>,
+    pub rating_tmdb: Option<f32>,
+    pub rating_imdb: Option<f32>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
@@ -39,20 +50,33 @@ pub struct CreateMovieEntry {
     pub is_primary: bool,
 }
 
+#[cfg(feature = "entity")]
 impl From<beam_entity::movie::Model> for Movie {
     fn from(model: beam_entity::movie::Model) -> Self {
         Self {
             id: model.id,
             title: model.title,
+            title_localized: model.title_localized,
+            description: model.description,
+            year: model.year.map(|y| y as u32),
+            release_date: model.release_date,
             runtime: model
                 .runtime_mins
                 .map(|mins| Duration::from_secs((mins * 60) as u64)),
+            poster_url: model.poster_url,
+            backdrop_url: model.backdrop_url,
+            tmdb_id: model.tmdb_id.map(|id| id as u32),
+            imdb_id: model.imdb_id,
+            tvdb_id: model.tvdb_id.map(|id| id as u32),
+            rating_tmdb: model.rating_tmdb,
+            rating_imdb: model.rating_imdb,
             created_at: model.created_at.with_timezone(&Utc),
             updated_at: model.updated_at.with_timezone(&Utc),
         }
     }
 }
 
+#[cfg(feature = "entity")]
 impl From<beam_entity::movie_entry::Model> for MovieEntry {
     fn from(model: beam_entity::movie_entry::Model) -> Self {
         Self {
