+++ beam-domain/src/repositories/mod.rs
@@ -1,0 +1,13 @@
+pub mod admin_log;
+pub mod file;
+pub mod library;
+pub mod movie;
+pub mod show;
+pub mod stream;
+
+pub use admin_log::AdminLogRepository;
+pub use file::FileRepository;
+pub use library::LibraryRepository;
+pub use movie::MovieRepository;
+pub use show::ShowRepository;
+pub use stream::MediaStreamRepository;
