+++ beam-index/src/repositories/stream.rs
@@ -2,15 +2,17 @@ use async_trait::async_trait;
 use sea_orm::{DatabaseConnection, DbErr};
 use uuid::Uuid;
 
-use crate::models::domain::{CreateMediaStream, MediaStream};
+pub use beam_domain::repositories::MediaStreamRepository;
 
-/// Repository for managing media stream persistence operations.
-#[cfg_attr(any(test, feature = "test-utils"), mockall::automock)]
-#[async_trait]
-pub trait MediaStreamRepository: Send + Sync + std::fmt::Debug {
-    async fn insert_streams(&self, streams: Vec<CreateMediaStream>) -> Result<u32, DbErr>;
-    async fn find_by_file_id(&self, file_id: Uuid) -> Result<Vec<MediaStream>, DbErr>;
+#[cfg(any(test, feature = "test-utils"))]
+pub use beam_domain::repositories::stream::MockMediaStreamRepository;
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod in_memory {
+    pub use beam_domain::repositories::stream::in_memory::*;
 }
+
+use crate::models::domain::{CreateMediaStream, MediaStream};
 
 /// SQL-based implementation of the MediaStreamRepository trait.
 #[derive(Debug, Clone)]
@@ -146,54 +148,5 @@ impl MediaStreamRepository for SqlMediaStreamRepository {
             .await?;
 
         Ok(models.into_iter().map(MediaStream::from).collect())
-    }
-}
-
-/// In-memory implementation for use in tests and test-utils consumers.
-#[cfg(any(test, feature = "test-utils"))]
-pub mod in_memory {
-    use super::*;
-    use std::collections::HashMap;
-    use std::sync::Mutex;
-
-    #[derive(Debug, Default)]
-    pub struct InMemoryMediaStreamRepository {
-        pub streams: Mutex<HashMap<Uuid, Vec<MediaStream>>>,
-    }
-
-    #[async_trait]
-    impl MediaStreamRepository for InMemoryMediaStreamRepository {
-        async fn insert_streams(&self, streams: Vec<CreateMediaStream>) -> Result<u32, DbErr> {
-            let count = streams.len() as u32;
-            for create in streams {
-                let stream = MediaStream {
-                    id: Uuid::new_v4(),
-                    file_id: create.file_id,
-                    index: create.index,
-                    stream_type: create.stream_type,
-                    codec: create.codec,
-                    metadata: create.metadata,
-                };
-                self.streams
-                    .lock()
-                    .unwrap()
-                    .entry(create.file_id)
-                    .or_default()
-                    .push(stream);
-            }
-            Ok(count)
-        }
-
-        async fn find_by_file_id(&self, file_id: Uuid) -> Result<Vec<MediaStream>, DbErr> {
-            let mut streams = self
-                .streams
-                .lock()
-                .unwrap()
-                .get(&file_id)
-                .cloned()
-                .unwrap_or_default();
-            streams.sort_by_key(|s| s.index);
-            Ok(streams)
-        }
     }
 }
