+++ beam-stream/Cargo.toml
@@ -8,6 +8,7 @@ default-run = "beam-stream"
 
 [dependencies]
 beam-auth = { path = "../beam-auth", features = ["utils"] }
+beam-domain = { path = "../beam-domain" }
 beam-entity = { path = "../beam-entity" }
 beam-index = { path = "../beam-index" }
 http = { workspace = true }
@@ -61,9 +62,10 @@ xxhash-rust = { version = "0.8.15", features = ["xxh3"] }
 [dev-dependencies]
 futures = "0.3.31"
 beam-auth = { path = "../beam-auth", features = ["test-utils"] }
+beam-domain = { path = "../beam-domain", features = ["test-utils"] }
 beam-migration = { path = "../beam-migration" }
 beam-index = { path = "../beam-index", features = ["test-utils"] }
 mockall = "0.14.0"
 
 [features]
-test-utils = ["beam-index/test-utils"]
+test-utils = ["beam-domain/test-utils", "beam-index/test-utils"]
