--- beam-stream/src/repositories/file.rs
@@ -1,207 +1,0 @@
-use async_trait::async_trait;
-use sea_orm::{DatabaseConnection, DbErr};
-
-use crate::models::domain::{CreateMediaFile, MediaFile, UpdateMediaFile};
-use uuid::Uuid;
-
-/// Repository for managing media file persistence operations.
-///
-/// This trait defines the data access layer for media files (videos) within libraries.
-/// It handles file metadata storage including paths, hashes, sizes, and associations
-/// with movies or TV episodes.
-#[cfg_attr(test, mockall::automock)]
-#[async_trait]
-pub trait FileRepository: Send + Sync + std::fmt::Debug {
-    /// Finds a media file by its filesystem path.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - The filesystem path to search for
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(Some(MediaFile))` - If a file with the given path exists
-    /// * `Ok(None)` - If no file with the given path exists
-    /// * `Err(DbErr)` - If a database error occurs
-    async fn find_by_path(&self, path: &str) -> Result<Option<MediaFile>, DbErr>;
-
-    /// Finds all media files belonging to a specific library.
-    async fn find_all_by_library(&self, library_id: Uuid) -> Result<Vec<MediaFile>, DbErr>;
-
-    /// Creates a new media file record in the database.
-    ///
-    /// # Arguments
-    ///
-    /// * `create` - The parameters for creating the media file
-    ///
-    /// # Returns
-    ///
-    /// The newly created media file with its generated ID and timestamps,
-    /// or a database error if creation fails.
-    async fn create(&self, create: CreateMediaFile) -> Result<MediaFile, DbErr>;
-
-    /// Updates an existing media file record.
-    async fn update(&self, update: UpdateMediaFile) -> Result<MediaFile, DbErr>;
-
-    /// Deletes a media file by its ID.
-    async fn delete(&self, id: Uuid) -> Result<(), DbErr>;
-
-    /// Deletes multiple media files by their IDs.
-    async fn delete_by_ids(&self, ids: Vec<Uuid>) -> Result<u64, DbErr>;
-}
-
-/// SQL-based implementation of the FileRepository trait.
-///
-/// Uses SeaORM to interact with a relational database (PostgreSQL).
-#[derive(Debug, Clone)]
-pub struct SqlFileRepository {
-    db: DatabaseConnection,
-}
-
-impl SqlFileRepository {
-    /// Creates a new SQL file repository.
-    ///
-    /// # Arguments
-    ///
-    /// * `db` - The database connection to use for all operations
-    pub fn new(db: DatabaseConnection) -> Self {
-        Self { db }
-    }
-}
-
-#[async_trait]
-impl FileRepository for SqlFileRepository {
-    async fn find_by_path(&self, path: &str) -> Result<Option<MediaFile>, DbErr> {
-        use beam_entity::files;
-        use sea_orm::{ColumnTrait, EntityTrait, QueryFilter};
-
-        let model = files::Entity::find()
-            .filter(files::Column::FilePath.eq(path))
-            .one(&self.db)
-            .await?;
-
-        Ok(model.map(MediaFile::from))
-    }
-
-    async fn find_all_by_library(&self, library_id: Uuid) -> Result<Vec<MediaFile>, DbErr> {
-        use beam_entity::files;
-        use sea_orm::{ColumnTrait, EntityTrait, QueryFilter};
-
-        let models = files::Entity::find()
-            .filter(files::Column::LibraryId.eq(library_id))
-            .all(&self.db)
-            .await?;
-
-        Ok(models.into_iter().map(MediaFile::from).collect())
-    }
-
-    async fn create(&self, create: CreateMediaFile) -> Result<MediaFile, DbErr> {
-        use crate::models::domain::MediaFileContent;
-        use beam_entity::files;
-        use chrono::Utc;
-        use sea_orm::{ActiveModelTrait, Set};
-
-        let now = Utc::now();
-        let (movie_entry_id, episode_id) = match create.content {
-            Some(MediaFileContent::Movie { movie_entry_id }) => (Some(movie_entry_id), None),
-            Some(MediaFileContent::Episode { episode_id }) => (None, Some(episode_id)),
-            None => (None, None),
-        };
-
-        let new_file = files::ActiveModel {
-            id: Set(uuid::Uuid::new_v4()),
-            library_id: Set(create.library_id),
-            file_path: Set(create.path.to_string_lossy().to_string()),
-            hash_xxh3: Set(create.hash as i64),
-            file_size: Set(create.size_bytes as i64),
-            mime_type: Set(create.mime_type),
-            duration_secs: Set(create.duration.map(|d| d.as_secs_f64())),
-            container_format: Set(create.container_format),
-            language: Set(None),
-            quality: Set(None),
-            release_group: Set(None),
-            is_primary: Set(true),
-            movie_entry_id: Set(movie_entry_id),
-            episode_id: Set(episode_id),
-            scanned_at: Set(now.into()),
-            updated_at: Set(now.into()),
-            file_status: Set(create.status.to_string()),
-        };
-
-        let result = new_file.insert(&self.db).await?;
-        Ok(MediaFile::from(result))
-    }
-    async fn update(&self, update: UpdateMediaFile) -> Result<MediaFile, DbErr> {
-        use crate::models::domain::MediaFileContent;
-        use beam_entity::files;
-        use sea_orm::{ActiveModelTrait, Set};
-
-        let mut active_model: files::ActiveModel = files::ActiveModel {
-            id: Set(update.id),
-            ..Default::default()
-        };
-
-        if let Some(hash) = update.hash {
-            active_model.hash_xxh3 = Set(hash as i64);
-        }
-        if let Some(size) = update.size_bytes {
-            active_model.file_size = Set(size as i64);
-        }
-        if let Some(mime_type) = update.mime_type {
-            active_model.mime_type = Set(Some(mime_type));
-        }
-        if let Some(duration) = update.duration {
-            active_model.duration_secs = Set(Some(duration.as_secs_f64()));
-        }
-        if let Some(container) = update.container_format {
-            active_model.container_format = Set(Some(container));
-        }
-        if let Some(status) = update.status {
-            active_model.file_status = Set(status.to_string());
-        }
-
-        // Handle content update
-        if let Some(content) = update.content {
-            match content {
-                MediaFileContent::Movie { movie_entry_id } => {
-                    active_model.movie_entry_id = Set(Some(movie_entry_id));
-                    active_model.episode_id = Set(None);
-                }
-                MediaFileContent::Episode { episode_id } => {
-                    active_model.movie_entry_id = Set(None);
-                    active_model.episode_id = Set(Some(episode_id));
-                }
-            }
-        }
-
-        // Also update timestamp
-        active_model.updated_at = Set(chrono::Utc::now().into());
-
-        let result = active_model.update(&self.db).await?;
-        Ok(MediaFile::from(result))
-    }
-
-    async fn delete(&self, id: Uuid) -> Result<(), DbErr> {
-        use beam_entity::files;
-        use sea_orm::EntityTrait;
-
-        files::Entity::delete_by_id(id).exec(&self.db).await?;
-        Ok(())
-    }
-
-    async fn delete_by_ids(&self, ids: Vec<Uuid>) -> Result<u64, DbErr> {
-        use beam_entity::files;
-        use sea_orm::{ColumnTrait, EntityTrait, QueryFilter};
-
-        if ids.is_empty() {
-            return Ok(0);
-        }
-
-        let result = files::Entity::delete_many()
-            .filter(files::Column::Id.is_in(ids))
-            .exec(&self.db)
-            .await?;
-
-        Ok(result.rows_affected)
-    }
-}
